… src/coreclr/gc/gc.h
… src/coreclr/gc/gcpriv.h
… src/coreclr/inc/gcmsg.inl